Generate Full URL for Route
Fetching the URI of the route is easy, forming a full URL for it is not
as easy. I wanted a way to build a full URL that used the "host" defined
as a route option instead of just using either the current host or just
the URI. This means you can now link to other sites that use the same
installation of ProtonSite or even sub domains.

+ Added url() method to Route class.

Read the wiki for more information on how to use this method/function.